+++ modules/openlmis-web/src/test/java/org/openlmis/web/controller/ProgramControllerTest.java
@@ -63,10 +63,10 @@ public class ProgramControllerTest {
         when(session.getAttribute("USER")).thenReturn("dummyUser");
         request.setSession(session);
 
-        RoleAssignment roleAssignment = new RoleAssignment(1L, 2L, program.getId());
+        RoleAssignment roleAssignment = new RoleAssignment(1, 2, program.getId());
         List<RoleAssignment> roleAssignments = new ArrayList<>(Arrays.asList(roleAssignment));
         when(roleRightsService.getProgramWithGivenRightForAUser(Right.CREATE_REQUISITION, "dummyUser")).thenReturn(roleAssignments);
-        Long facilityId = 12345L;
+        Integer facilityId = 12345;
         when(programService.filterActiveProgramsAndFacility(roleAssignments, facilityId)).thenReturn(programs);
 
         assertEquals(programs, controller.getUserSupportedProgramsToCreateRequisition(facilityId, request));
